Ho sistemato le classi per le segnalazioni e i contest, ho modificato le entità utente, contenuto e PuntoDiInteresse rendendole normali invece che astratte per non avere problemi con le repository, ho aggiunto i joincolumn alla classe TuristaAutenticato, Contenuto per gestire i set nelle classi service, ho aggiunto le classi contenuticontroller, contenutistati, statosegnalazioni e le classi mapservice mapcontroller e map.html per openstreetmap. Vanno sistemate le classi per i comuni per gestire i contenuti usando joincolumn con le annotazioni @Table e @Data, le colonne devono avere lo stesso nome delle table aggiungendo _id per l'id

+++ src/main/java/com/it/unicam/progetto_ids_2023/service/ContenutiService.java
@@ -1,44 +1,122 @@
 package com.it.unicam.progetto_ids_2023.service;
 
-import com.it.unicam.progetto_ids_2023.model.contenuto.ContenutoMultimediale;
-import com.it.unicam.progetto_ids_2023.model.contenuto.ContenutoTestuale;
+import com.it.unicam.progetto_ids_2023.model.contenuto.*;
+import com.it.unicam.progetto_ids_2023.model.puntodiinteresse.Comune;
+import com.it.unicam.progetto_ids_2023.model.puntodiinteresse.PuntoDiInteresse;
 import com.it.unicam.progetto_ids_2023.model.utente.Contributor;
 import com.it.unicam.progetto_ids_2023.model.utente.ContributorAutorizzato;
 import com.it.unicam.progetto_ids_2023.model.utente.Curatore;
 import com.it.unicam.progetto_ids_2023.model.utente.Utente;
 import com.it.unicam.progetto_ids_2023.repository.ContenutoMultimedialeRepository;
 import com.it.unicam.progetto_ids_2023.repository.ContenutoTestualeRepository;
+import org.springframework.stereotype.Service;
+
+
+@Service
 
 public class ContenutiService {
 
     private ContenutoMultimedialeRepository multiRepo;
     private ContenutoTestualeRepository testoRepo;
 
-    public void addContenuto(Utente utente, ContenutoMultimediale multimediale, ContenutoTestuale testuale){
+    public void addContenuto(Utente utente, ContenutoMultimediale multimediale, ContenutoTestuale testuale, PuntoDiInteresse puntoDiInteresse, Comune comune){
 
         if(utente instanceof ContributorAutorizzato || utente instanceof Curatore) {
+       /*  // multimediale.setUtente(utente);
+            multimediale.setComune(comune);
+            multimediale.setPuntoDiInteresse(puntoDiInteresse);
+          //  testuale.setUtente(utente);
+            testuale.setComune(comune);
+            testuale.setPuntoDiInteresse(puntoDiInteresse);
+
+
             multimediale.setPending(false);
             testuale.setPending(false);
             multiRepo.save(multimediale);
-            testoRepo.save(testuale);
+            testoRepo.save(testuale);*/
         } else if (utente instanceof Contributor){
+      //     multimediale.setUtente(utente);
+  /*          multimediale.setComune(comune);
+            multimediale.setPuntoDiInteresse(puntoDiInteresse);
+        //    testuale.setUtente(utente);
+            testuale.setComune(comune);
+            testuale.setPuntoDiInteresse(puntoDiInteresse);
+
             multimediale.setPending(true);
             testuale.setPending(true);
             multiRepo.save(multimediale);
-            testoRepo.save(testuale);
+            testoRepo.save(testuale); */
 
         } else {
             throw new IllegalArgumentException();
         }
 
+
+
     }
 
-    public void deleteContenuto(Long multiMediaid, Long testoId){
+    public void deleteContenutoTestuale(Long multiMediaid, Long testoId){
         ContenutoMultimediale multimedia = multiRepo.findById(multiMediaid).orElseThrow();
         ContenutoTestuale testuale = testoRepo.findById(testoId).orElseThrow();
+        if(testuale.getStati().equals(ContenutiStati.RIFIUTATO)) {
 
-        multiRepo.delete(multimedia);
-        testoRepo.delete(testuale);
+            testoRepo.delete(testuale);
+        } else {
+            throw new IllegalArgumentException();
+        }
     }
 
+    public void deleteContenutoMultimediale(Long contenutoMediaId){
+       /* ContenutoMultimediale multimedia = multiRepo.findById(multiMediaid).orElseThrow();
+        if(testuale.getStati().equals(ContenutiStati.RIFIUTATO)) {
+
+            testoRepo.delete(testuale);
+        } else {
+            throw new IllegalArgumentException();
+        } */
+
+
+    }
+
+
+
+
+    public void rifiutaContenutoTestuale(Long contenutoTestoId){
+
+        ContenutoTestuale contenutoTestuale = testoRepo.findById(contenutoTestoId).orElseThrow();
+        if(contenutoTestuale.isPending()){
+            contenutoTestuale.setStato(ContenutiStati.RIFIUTATO);
+            testoRepo.save(contenutoTestuale);
+        }
+        else {
+            throw new IllegalArgumentException();
+        }
+
+
+    }
+
+    public void rifiutaContenutoMultimediale(Long contenutoTestoId){
+
+        ContenutoTestuale contenutoTestuale = testoRepo.findById(contenutoTestoId).orElseThrow();
+        if(contenutoTestuale.isPending()){
+            contenutoTestuale.setStato(ContenutiStati.RIFIUTATO);
+            testoRepo.save(contenutoTestuale);
+        }
+        else {
+            throw new IllegalArgumentException();
+        }
+
+
+    }
+
+
+
 }
+
+
+
+
+
+
+
+
